BackEndDjango/requirements.txt
# Django Backend Dependencies
Django>=4.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
python-dotenv>=1.0
